+++ CHANGELOG.md
@@ -62,6 +62,7 @@
 * `PUnsafeLiftDecl` and `PConstantDecl` as they are replaced by `PLiftable`
 * `Plutarch` module, as it served no useful purpose and was just confusing
 * `PType` synonym (use `S -> Type` honestly instead)
+* `PSBool` and functionality (now in `plutarch-ledger-api`)
 
 ### Fixed